requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
tqdm>=4.66.1
torch>=2.1.0
